플레이어 이동 데드락커닝 적용 시뮬레이션 적용 예정
