+++ Assets/Scripts/Player/OtherPlayer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2b24c3ec35b9a242980dacb8465ea1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
